implement sql, first pass
in the first pass, decided to just use query params to set the
comparsion type.

So FilterType set in schema is skipped for now. Instead, Constraint in
query has been expanded to include string matches and in array for now.
Any query that attempts to use these operators will be passed through to
db. Future work is to prevent those when generating query_ir, and to
allow parsing according to schema-defined filter type, instead of by
matching on the string.

The query ir generated in the future should be the same basically.
Because the info needed to generate sql is the same. There will just be
a different check and parsing before it.

For query_ir, add an `is_text` field.

In preparation for future, FilterType in schema's `Numeric` value is
changed to `Compare`