First pass parsing filters at query_ir stage with schema input
Previously, filters were parsed between ApiQueryOpt (the query params of
the url) and the Query. However, parsing this way means that the schema
has no input into the parsing, and the operator must be specified in the
query.

Now, in order to streamline the user interface, the operator is
specified in the schema (for compare, string match, in arrary), so that
it doesn't have to be specified in the url. To do that, the parsing is
also moved back to the query_ir stage, instead of the query stage.

Also, moved shared functionality for api to own module.

Also, cleaned up imports.